Fix #83 Make actual calculation for rant decay function.

+++ src/main/java/com/kizy/data/rant/Rant.java
@@ -25,7 +25,7 @@ public interface Rant {
 
     int getRantPower();
 
-    void changePower(int amount);
+    void changePower(double amount);
 
     boolean isAlive();
